consumer-backend/api/requirements.txt
starsessions~=2.1.3
fastapi~=0.114.1
starlette~=0.38.5
Authlib~=1.3.2
pydantic~=2.9.1
uvicorn~=0.30.6
SQLAlchemy~=2.0.32
httpx~=0.27.2
redis~=5.0.7
psycopg2-binary~=2.9.9
sqlmodel~=0.0.22
asyncpg~=0.29.0
fastapi-async-sqlalchemy~=0.6.1
greenlet~=3.1.1
pyjwt>=2.9.0
debugpy>=1.8.6
dateutils>=0.6.12